refactor(fit): dedupe FIT time helper and drop pass-through re-exports
- messages/activity.mapper: reuse convertFitTimeCreatedToIso from the
  health metadata builder instead of a byte-identical local copy
- record: inline the record.mapper re-export barrel into the two
  converters (mirrors the lap adapter, which has no such barrel)
- session/index: remove the unused convertKrdToFitSession and
  mapKrdSessionToFit re-exports (only reached directly, never via barrel)

No behavior change; coverage and all 462 tests unchanged.
