+++ .changeset/fit-code-reduction.md
@@ -1,0 +1,5 @@
+---
+"@kaiord/fit": patch
+---
+
+Internal code reduction, no behavior change.
