+++ packages/fit/src/adapters/record/fit-to-krd-record.converter.ts
@@ -2,7 +2,7 @@ import type { KRDRecord } from "@kaiord/core";
 
 import { type FitRecord, fitRecordSchema } from "../schemas/fit-record";
 import { validateCoordinates } from "../shared/coordinate.converter";
-import { mapFitRecordToKrd } from "./record.mapper";
+import { mapFitRecordToKrd } from "./record-from-fit.mapper";
 
 /**
  * Converts a FIT RECORD message to KRD record format.
