+++ packages/fit/src/adapters/record/krd-to-fit-record.converter.ts
@@ -1,7 +1,7 @@
 import { type KRDRecord, krdRecordSchema } from "@kaiord/core";
 
 import type { FitRecord } from "../schemas/fit-record";
-import { mapKrdRecordToFit } from "./record.mapper";
+import { mapKrdRecordToFit } from "./record-to-fit.mapper";
 
 /**
  * Converts a KRD record to FIT RECORD message format.
