--- packages/fit/src/adapters/record/record.mapper.ts
@@ -1,2 +1,0 @@
-export { mapFitRecordToKrd } from "./record-from-fit.mapper";
-export { mapKrdRecordToFit } from "./record-to-fit.mapper";
