+++ packages/fit/src/adapters/session/index.ts
@@ -1,3 +1,2 @@
 export { convertFitToKrdSession } from "./fit-to-krd-session.converter";
-export { convertKrdToFitSession } from "./krd-to-fit-session.converter";
-export { mapFitSessionToKrd, mapKrdSessionToFit } from "./session.mapper";
+export { mapFitSessionToKrd } from "./session.mapper";
